move some functions to new scrapy_poet.utils.testing module
