+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from scrapy_poet.utils import create_scrapy_settings
+from scrapy_poet.utils.testing import create_scrapy_settings
 
 
 @pytest.fixture()
