+++ tests/test_downloader.py
@@ -15,15 +15,15 @@ from web_poet.exceptions import HttpError, HttpRequestError, HttpResponseError
 from web_poet.pages import WebPage
 
 from scrapy_poet.downloader import create_scrapy_downloader
-from scrapy_poet.utils import (
+from scrapy_poet.utils import http_request_to_scrapy_request
+from scrapy_poet.utils.mockserver import MockServer
+from scrapy_poet.utils.testing import (
     AsyncMock,
     DelayedResource,
     EchoResource,
     StatusResource,
-    http_request_to_scrapy_request,
     make_crawler,
 )
-from scrapy_poet.utils.mockserver import MockServer
 
 
 @pytest.fixture
